requirements.txt
google-genai
fal-client
pyyaml
Pillow
requests
python-dotenv
